fix: bug 2249 pick invalid action when model untrained
